+++ sdk/communication/azure-communication-messages/src/samples/java/com/azure/communication/messages/ReadmeSamples.java
@@ -4,7 +4,18 @@
 
 package com.azure.communication.messages;
 
-import com.azure.communication.messages.models.*;
+import com.azure.communication.messages.models.AudioNotificationContent;
+import com.azure.communication.messages.models.DocumentNotificationContent;
+import com.azure.communication.messages.models.ImageNotificationContent;
+import com.azure.communication.messages.models.MessageTemplate;
+import com.azure.communication.messages.models.MessageTemplateBindings;
+import com.azure.communication.messages.models.MessageTemplateItem;
+import com.azure.communication.messages.models.MessageTemplateText;
+import com.azure.communication.messages.models.MessageTemplateValue;
+import com.azure.communication.messages.models.SendMessageResult;
+import com.azure.communication.messages.models.TemplateNotificationContent;
+import com.azure.communication.messages.models.TextNotificationContent;
+import com.azure.communication.messages.models.VideoNotificationContent;
 import com.azure.communication.messages.models.channels.WhatsAppMessageTemplateBindings;
 import com.azure.communication.messages.models.channels.WhatsAppMessageTemplateBindingsComponent;
 import com.azure.communication.messages.models.channels.WhatsAppMessageTemplateItem;
@@ -168,7 +179,7 @@ public final class ReadmeSamples {
     // BEGIN: readme-sample-sendVideoMessage
     /*
      * This sample shows how to send video message with below details
-     * Supported video - video/3gp (.3gp), 	video/mp4 (.mp4)
+     * Supported video - video/3gp (.3gp), video/mp4 (.mp4)
      * Note: Business cannot initiate conversation with media message.
      * */
     public void sendVideoMessage() {