+++ sdk/communication/azure-communication-messages/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Release History
 
-## 1.1.0-beta.1 (Unreleased)
+## 1.1.0 (2024-10-30)
 
 ### Features Added
+- Added ImageNotificationContent to send image messgae.
+- Added DocumentNotificationContent to send document message.
+- Added VideoNotificationContent to send video message.
+- Added AudioNotificationContent to  send audio message.
+- Deprecated MediaNotificationContent.
 
 ### Breaking Changes